Booking finish

+++ Models/Property.cs
@@ -39,7 +39,7 @@ namespace School.OnlineBookingSystem.Models
         [DataMember(EmitDefaultValue = false)]
         public Dictionary<string, string> ApartmentDic { get; set; }
         
-        public float MinPrice { get; set; }
+        public decimal MinPrice { get; set; }
 
         #endregion
